Refactor to utilize new dynamic shaping of tensor

+++ include/nn/layers_impl/activation_layer.tpp
@@ -98,13 +98,13 @@ uint64_t ActivationLayer<T>::backward_flops(const std::vector<size_t> &input_sha
 template <typename T>
 uint64_t ActivationLayer<T>::forward_complexity(const std::vector<size_t> &input_shape) const {
   return static_cast<uint64_t>(
-      std::min(forward_flops(input_shape), static_cast<uint64_t>(UINT32_MAX)));
+      std::min(forward_flops(input_shape), static_cast<uint64_t>(UINT64_MAX)));
 }
 
 template <typename T>
 uint64_t ActivationLayer<T>::backward_complexity(const std::vector<size_t> &input_shape) const {
   return static_cast<uint64_t>(
-      std::min(backward_flops(input_shape), static_cast<uint64_t>(UINT32_MAX)));
+      std::min(backward_flops(input_shape), static_cast<uint64_t>(UINT64_MAX)));
 }
 
 template class ActivationLayer<float>;
