+++ include/nn/layers_impl/dropout_layer.tpp
@@ -164,13 +164,13 @@ template <typename T>
 uint64_t DropoutLayer<T>::forward_complexity(const std::vector<size_t> &input_shape) const {
 
   return static_cast<uint64_t>(
-      std::min(forward_flops(input_shape), static_cast<uint64_t>(UINT32_MAX)));
+      std::min(forward_flops(input_shape), static_cast<uint64_t>(UINT64_MAX)));
 }
 
 template <typename T>
 uint64_t DropoutLayer<T>::backward_complexity(const std::vector<size_t> &input_shape) const {
   return static_cast<uint64_t>(
-      std::min(backward_flops(input_shape), static_cast<uint64_t>(UINT32_MAX)));
+      std::min(backward_flops(input_shape), static_cast<uint64_t>(UINT64_MAX)));
 }
 
 template class DropoutLayer<float>;
