+++ include/nn/layers_impl/slice_layer.tpp
@@ -131,13 +131,13 @@ uint64_t SliceLayer<T>::backward_flops(const std::vector<size_t> &input_shape) c
 template <typename T>
 uint64_t SliceLayer<T>::forward_complexity(const std::vector<size_t> &input_shape) const {
   return static_cast<uint64_t>(
-      std::min(forward_flops(input_shape), static_cast<uint64_t>(UINT32_MAX)));
+      std::min(forward_flops(input_shape), static_cast<uint64_t>(UINT64_MAX)));
 }
 
 template <typename T>
 uint64_t SliceLayer<T>::backward_complexity(const std::vector<size_t> &input_shape) const {
   return static_cast<uint64_t>(
-      std::min(backward_flops(input_shape), static_cast<uint64_t>(UINT32_MAX)));
+      std::min(backward_flops(input_shape), static_cast<uint64_t>(UINT64_MAX)));
 }
 
 template class SliceLayer<float>;
